+++ install.ps1
@@ -130,7 +130,9 @@ $RedactMatcher  = 'Bash|Read|Grep'
 $NoticeMatcher  = 'Edit|Write|mcp__.*'
 $FailureMatcher = 'Bash|Read|Grep|Edit|Write|mcp__.*'
 
-$Readers = @('cat','head','tail','less','more','strings','xxd','od','nl','tac','bat','batcat')
+$Readers = @('cat','head','tail','less','more','strings','xxd','od','nl','tac','bat','batcat',
+             'grep','egrep','fgrep','rg','ag','ack','sed','awk','gawk','mawk','sort','uniq',
+             'cut','rev','column','jq','yq')
 $SecretFiles = @('*.env','*.env.*','*/.env','*.pem','*.key','*.p12','*.pfx',
                  '*id_rsa*','*id_ed25519*','*id_ecdsa*','*/.bashrc*','*/.zshrc*',
                  '*/.profile*','*/.bash_profile*','*/.zshenv*','*/.netrc*','*_netrc*',